# docs/swagger.yaml
definitions:
...
  controllers.SearchResult:
    properties:
      activities:
        items:
          $ref: '#/definitions/models.Activity'
        type: array
      address:
        $ref: '#/definitions/models.Address'
      associated_services:
        items:
          $ref: '#/definitions/models.Service'
        type: array
      working_area:
        $ref: '#/definitions/models.WorkingArea'
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Activity:
    properties:
      activityName:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/models.ActivityItem'
        type: array
      kycID:
        type: integer
      updatedAt:
        type: string
      workingAreaID:
        type: integer
    type: object
  models.ActivityItem:
    properties:
      activityID:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  models.Address:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      district:
        type: string
      id:
        type: integer
      kycID:
        type: integer
      latitude:
        type: string
      longitude:
        type: string
      municipality:
        type: string
      province:
        type: string
      updatedAt:
        type: string
      userID:
        type: integer
      wardNumber:
        type: string
    type: object
  models.InvestmentOption:
    enum:
    - up to 5 Lakhs
    - up to 10 Lakhs
    - up to 25 Lakhs
    - up to 50 Lakhs
    - up to 1 Crore
    - above 1 Crore
    type: string
    x-enum-varnames:
    - UpTo5LAKHS
    - UpTo10LAKHS
    - UpTo25LAKHS
    - UpTo50LAKHS
    - UpTo1CRORE
    - Above1CRORE
  models.Kyc:
    properties:
      address:
        $ref: '#/definitions/models.Address'
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      firmRegistered:
        type: boolean
      fullName:
        type: string
      id:
        type: integer
      mobileNumber:
        type: string
      service:
        $ref: '#/definitions/models.Service'
      updatedAt:
        type: string
      userID:
        type: integer
      workingArea:
        $ref: '#/definitions/models.WorkingArea'
    type: object
  models.Service:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      investment:
        $ref: '#/definitions/models.InvestmentOption'
      kycID:
        type: integer
      serviceName:
        $ref: '#/definitions/models.ServiceType'
      updatedAt:
        type: string
      userID:
        type: integer
    type: object
  models.ServiceType:
    enum:
    - Expert Advice
    - Business Partnership
    - Bank Loan Facilitation
    - Training and Coaching
    - Cold Store Construction
    - Assistance in Marketing
    - Investment
    type: string
    x-enum-varnames:
    - ExpertAdvice
    - BusinessPartnership
    - BankLoanFacilitation
    - TrainingAndCoaching
    - ColdStoreConstruction
    - AssistanceInMarketing
    - InvestmentService
  models.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      is_kyc:
        type: boolean
      kyc:
        $ref: '#/definitions/models.Kyc'
      password:
        type: string
      profile_picture:
        type: string
      role:
        type: string
      updatedAt:
        type: string
    required:
    - email
    - password
    type: object
  models.WorkingArea:
    properties:
      activities:
        items:
          $ref: '#/definitions/models.Activity'
        type: array
      areaName:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      kycID:
        type: integer
      updatedAt:
        type: string
      userID:
        type: integer
    type: object
info:
  contact: {}
paths:
  /search:
    get:
      consumes:
      - application/json
      description: Search for users by email
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of users matching the search query
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      summary: Perform a search
      tags:
      - Search
  /search/address:
    get:
      consumes:
      - application/json
      description: Search based on the address model and return associated working
        area, activities, and services
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Search result
          schema:
            $ref: '#/definitions/controllers.SearchResult'
      summary: Perform a search based on the address model
      tags:
      - Search
  /search/advanced:
    get:
      consumes:
      - application/json
      description: Perform an advanced global search across all models based on the
        provided query string
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of results matching the advanced search query
          schema:
            items:
              $ref: '#/definitions/models.Kyc'
            type: array
      summary: Perform an advanced global search across all models
      tags:
      - Search
  /search/all:
    get:
      consumes:
      - application/json
      description: Perform a global search across all models
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Perform a global search across all models
      tags:
      - Search
  /search/all/address/{province}/{district}/{municipality}/{ward_number}:
    get:
      consumes:
      - application/json
      description: Search based on the address model and return associated working
        area, activities, and services
      parameters:
      - description: Province
        in: path
        name: province
        required: true
        type: string
      - description: District
        in: path
        name: district
        required: true
        type: string
      - description: Municipality
        in: path
        name: municipality
        required: true
        type: string
      - description: Ward Number
        in: path
        name: ward_number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Search result
          schema:
            $ref: '#/definitions/controllers.SearchResult'
      summary: Perform a search based on the address model
      tags:
      - Search
...
swagger: "2.0"